+++ packages/target-tauri/src-tauri/src/html_window.rs
@@ -1,23 +1,21 @@
 use std::{path::PathBuf, str::FromStr, sync::Arc};
 
-use log::{debug, error, info, warn};
-use serde::Serialize;
+use log::{error, info, warn};
+
 use tauri::{
-    async_runtime::block_on,
-    menu::{CheckMenuItem, Menu, MenuItem},
-    webview::WebviewBuilder,
-    LogicalPosition, LogicalSize, Manager, PhysicalPosition, Url, Webview, WebviewUrl, Window,
-    WindowBuilder, WindowEvent,
+    async_runtime::block_on, webview::WebviewBuilder, LogicalPosition, LogicalSize, Manager, Url,
+    Webview, WebviewUrl, Window, WindowBuilder, WindowEvent,
 };
 use tauri_plugin_dialog::DialogExt;
 use tauri_plugin_opener::OpenerExt;
 use tauri_plugin_store::StoreExt;
 
 use crate::{
+    html_window::error::Error,
+    html_window::punycode::{puny_code_decode_host, puny_code_encode_host},
     settings::{
         get_content_protection, get_setting_bool_or, CONFIG_FILE,
         HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_DEFAULT, HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_KEY,
-        HTML_EMAIL_WARNING_DEFAULT, HTML_EMAIL_WARNING_KEY,
     },
     state::html_email_instances::InnerHtmlEmailInstanceData,
     temp_file::get_temp_folder_path,
@@ -28,168 +26,10 @@ const HEADER_HEIGHT: f64 = 100.;
 const DEFAULT_WINDOW_WIDTH: f64 = 800.;
 const DEFAULT_WINDOW_HEIGHT: f64 = 600.;
 
-#[derive(Debug, thiserror::Error)]
-pub(crate) enum Error {
-    #[error(transparent)]
-    Tauri(#[from] tauri::Error),
-    // #[error(transparent)]
-    // Wry(#[from] wry::Error),
-    #[error("window label not found in HtmlEmailInstancesState")]
-    WindowNotFoundInState,
-    #[error(transparent)]
-    Store(#[from] tauri_plugin_store::Error),
-    #[error("user canceled")]
-    UserCanceled,
-    #[error(transparent)]
-    TokioOneshotRecv(#[from] tokio::sync::oneshot::error::RecvError),
-}
-
-impl serde::Serialize for Error {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::ser::Serializer,
-    {
-        serializer.serialize_str(self.to_string().as_ref())
-    }
-}
-
-async fn set_load_remote_content(webview: &tauri::Webview, new_state: bool) -> Result<(), Error> {
-    let instance_state = webview.state::<HtmlEmailInstancesState>();
-    instance_state
-        .set_network_allow_state(webview.window().label(), new_state)
-        .await;
-
-    // reload header & html email webviews to apply changes
-    for mut webview in webview.window().webviews() {
-        if webview.label().ends_with("-mail") {
-            // TODO: wry webview.reload is missing
-            // this is a hack to emulate reload
-            // IDEA: better error handling
-            let url = webview.url().unwrap();
-            webview
-                .navigate(Url::from_str("about:blank").unwrap())
-                .unwrap();
-            webview.navigate(url).unwrap();
-        }
-    }
-    Ok(())
-}
-
-#[tauri::command]
-pub(crate) async fn html_email_set_load_remote_content(
-    app: tauri::AppHandle,
-    webview: tauri::Webview,
-    load_remote_content: bool,
-) -> Result<(), Error> {
-    let desktop_settings = app.store(CONFIG_FILE)?;
-    let warning = get_setting_bool_or(
-        desktop_settings.get(HTML_EMAIL_WARNING_KEY),
-        HTML_EMAIL_WARNING_DEFAULT,
-    );
-
-    if warning && load_remote_content {
-        let (tx, rx) = tokio::sync::oneshot::channel::<bool>();
-
-        app.dialog()
-            .message("tx('load_remote_content_ask')")
-            .parent(&webview.window())
-            .buttons(tauri_plugin_dialog::MessageDialogButtons::OkCancelCustom(
-                // TODO use translation strings, as soon as translations are avaialble in rust backend
-                "Yes".to_owned(),
-                "No".to_owned(),
-            ))
-            .show(|answer| tx.send(answer).unwrap());
-        if !rx.await? {
-            return Err(Error::UserCanceled);
-        }
-    }
-
-    set_load_remote_content(&webview, load_remote_content).await
-}
-
-#[tauri::command]
-pub(crate) fn html_email_open_menu(
-    app: tauri::AppHandle,
-    webview: tauri::Webview,
-    html_instances_state: tauri::State<HtmlEmailInstancesState>,
-) -> Result<(), Error> {
-    let desktop_settings = app.store(CONFIG_FILE)?;
-    let always_load = get_setting_bool_or(
-        desktop_settings.get(HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_KEY),
-        HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_DEFAULT,
-    );
-    let warning = get_setting_bool_or(
-        desktop_settings.get(HTML_EMAIL_WARNING_KEY),
-        HTML_EMAIL_WARNING_DEFAULT,
-    );
-
-    let instance_state = block_on(html_instances_state.get(webview.window().label()))
-        .ok_or(Error::WindowNotFoundInState)?;
-
-    let always_load_remote_images = CheckMenuItem::new(
-        &app,
-        "tx('always_load_remote_images')",
-        true,
-        always_load,
-        None::<&str>,
-    )?;
-    let show_warning = CheckMenuItem::new(&app, "tx('show_warning')", true, warning, None::<&str>)?;
-
-    let menu = if instance_state.is_contact_request {
-        Menu::with_items(&app, &[&show_warning])
-    } else {
-        Menu::with_items(&app, &[&always_load_remote_images, &show_warning])
-    }?;
-
-    // treefit: I haven't found an easy way to get the title bar offset,
-    // so now we just spawn it at mouse position
-    webview.window().popup_menu(&menu)?;
-    webview.window().on_menu_event(move |window, event| {
-        // IDEA: better error handling
-        let desktop_settings = window.store(CONFIG_FILE).expect("config file not found");
-        if event.id() == always_load_remote_images.id() {
-            let new_always_load = !always_load;
-            desktop_settings.set(HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_KEY, new_always_load);
-            block_on(set_load_remote_content(&webview, new_always_load)).unwrap();
-
-            // TODO: implement reload in tauri
-            webview.eval("location.reload()").unwrap();
-        }
-        if event.id() == show_warning.id() {
-            desktop_settings.set(HTML_EMAIL_WARNING_KEY, !warning);
-        }
-    });
-    Ok(())
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct HtmlEmailInfo {
-    subject: String,
-    sender: String, // this is called "from" in electron edition
-    receive_time: String,
-    toggle_network: bool,
-    network_button_label_text: String,
-}
-
-#[tauri::command]
-pub(crate) fn get_html_window_info(
-    webview: tauri::Webview,
-    html_instances_state: tauri::State<HtmlEmailInstancesState>,
-) -> Result<HtmlEmailInfo, Error> {
-    let label = webview.window().label().to_owned();
-    // IDEA: can we make this function async without getting an error on html_instances_state
-    let instance =
-        block_on(html_instances_state.get(&label)).ok_or(Error::WindowNotFoundInState)?;
-
-    Ok(HtmlEmailInfo {
-        subject: instance.subject,
-        sender: instance.sender,
-        receive_time: instance.receive_time,
-        toggle_network: instance.network_allow_state,
-        network_button_label_text: format!("tx(\"load_remote_content\")"),
-    })
-}
+pub(crate) mod commands;
+// mod email_scheme;
+mod error;
+mod punycode;
 
 #[tauri::command]
 pub(crate) fn open_html_window(
@@ -488,49 +328,4 @@ fn truncate_text(text: &str, max_len: usize) -> String {
     } else {
         truncated
     }
-}
-
-/// Returns true if host string contains non ASCII characters
-fn is_puny(host: &str) -> bool {
-    for ch in host.chars() {
-        if !(ch.is_ascii_alphanumeric() || matches!(ch, '.' | '-')) {
-            return true;
-        }
-    }
-    false
-}
-
-/// Returns host as punycode encoded string
-fn puny_code_encode_host(host: &str) -> String {
-    host.to_owned()
-        .split('.')
-        .map(|sub| {
-            if is_puny(sub) {
-                format!(
-                    "xn--{}",
-                    unic_idna_punycode::encode_str(sub)
-                        .unwrap_or_else(|| "[punycode encode failed]".to_owned())
-                )
-            } else {
-                sub.to_owned()
-            }
-        })
-        .collect::<Vec<String>>()
-        .join(".")
-}
-
-/// Returns host as decoded string
-fn puny_code_decode_host(host: &str) -> String {
-    host.to_owned()
-        .split('.')
-        .map(|sub| {
-            if let Some(sub) = sub.strip_prefix("xn--") {
-                unic_idna_punycode::decode_to_string(sub)
-                    .unwrap_or_else(|| "[punycode decode failed]".to_owned())
-            } else {
-                sub.to_owned()
-            }
-        })
-        .collect::<Vec<String>>()
-        .join(".")
 }
