+++ packages/target-tauri/src-tauri/src/lib.rs
@@ -118,9 +118,9 @@ pub fn run() {
             settings::change_desktop_settings_apply_side_effects,
             help_window::open_help_window,
             html_window::open_html_window,
-            html_window::get_html_window_info,
-            html_window::html_email_open_menu,
-            html_window::html_email_set_load_remote_content,
+            html_window::commands::get_html_window_info,
+            html_window::commands::html_email_open_menu,
+            html_window::commands::html_email_set_load_remote_content,
         ])
         .register_asynchronous_uri_scheme_protocol("webxdc-icon", webxdc::webxdc_icon_protocol)
         .register_asynchronous_uri_scheme_protocol("dcblob", blobs::delta_blobs_protocol)
